New Generator API (#41)
Split `ethcontract-derive` into two crates, the new one being called `ethcontract-generate`. This new crate is a library instead of a proc macro and implements all the code generation logic. This allows `ethcontract` bindings to be generated as part of a build script instead of only as a proc macro. The advantages of using a build script is that it works better with RLS and Racer autocompletion at the cost of some ergonomics (for example, a build script is required).

### Test Plan

CI

### Commit History

* generate crate now builds

* derive now works with the generate crate

* cargo fmt