Confirmations (#105)
This PR unifies the confirmation API between all transaction (deployment and method calls) and eliminates the `send_and_confirm` method as now confirmation parameters are part of the transaction builder.

Note that this also addresses some issues with confirmation. A new case was created to upstream some of these issues to `web3` crate here: #104. Details of the issues with the current confirmation implementation are in the aforementioned issue.

This closes #95 

### Test Plan

New unit tests for the confirmation implementation and add waiting for confirmation with Ganache example. (Note that Rinkeby example currently waits for confirmation already). 

### Commit History

* started implemented confirmation

* implemented confirmations

* transaction confirmation options

* confirmation support from transactions

* adjust examples for new confirmations

* added support for polling

* Added a bunch of unit tests

* added polling unit test

* initial confirmation implementation

* todo

* update comment

* check transaction result when it is confirmed

* added unit tests for transactions with new confirmation API

* fix comment typo

* fix copy pasta

* ability to never timeout