[Events] New Log Filter API (#291)
This PR defines a new log stream/log filter API for querying past logs. This is the start of a refactor of the current event querying and streaming implementation in `ethcontract` so that we can have a `eth_getLogs` based log and event stream due to some issues we've encountered with using filters behind a load balancer.

The newly introduced stream is also "fault-tolerant" in that it can deal with a node being temporarily down by retrying the request it was on.

Note that this API is currently unused, but is the start of a larger refactor captured by: #290

### Test Plan

New unit tests.

Co-authored-by: Nicholas Rodrigues Lordello <nlordell@gmail.com>