Implement batching for View methods (#455)
Implements #355

This PR adds a new BatchCall component which can enqueue ViewMethods and allows executing all queued calls in one roundtrip to the Node. When adding a view method to the batch a future is returned, that can only resolve once the batch has been executed. 

This can potentially lead to deadlocks (e.g. when the future is awaited before the call batch in the same thread). I considered checking for resolution inside each call future and returning an error if the batch has not yet been resolved, however this would potentially prevent legitimate use cases where we await the individual call futures in one thread and execute the batch in a different one.

Alternatively we could make it so that adding a call returns nothing and waiting on the batch transport returns a list of all responses. However, I wasn't sure how this would work in case we want to enqueue multiple different calls with different signatures. With the proposed solution we keep the type-safety of existing view method calls, which I found convenient.

### Test Plan

Added a unit test and `examples/batch.rs ` for a demonstration on how to use it.