Estimate gas for non offline transactions (#472)
I looked into why we have to specify a hardcoded gas_limit of 8M in both gp-v1 and gp-v2 services (this supposedly negatively affects the time our txs stay in the mempool as some nodes might only consider a tx if there is still enough "remaining gas" in the block according to the tx's gas limit). 

The issue seemed to mostly affect the e2e tests (in fact on mainnet not using a hardcoded gas limit works fine and on xdai it works when connecting to a good node that supports proper gas estimates - e.g. https://dai.poa.network)

I found that we are estimating gas only for offline signed transactions (which is what we use on mainnet/xdai). Furthermore, if `None` is specified as a gas limit [some nodes](https://eth.wiki/json-rpc/API#eth_signtransaction) (e.g. ganache) set the gas limit to a hardcoded 90k, which is not enough for most of our transactions. This explains why our txs fail in the e2e tests where we are using local signing.

This PR suggests to run an estimate gas call also for local signing and locked account transactions. Alternatively we could do an explicit estimate_gas call inside our services and set the value appropriately. However this would have to happen on all callers.